requirements.txt
absl-py==0.4.1
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.15.0
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
matplotlib==2.2.3
mccabe==0.6.1
numpy==1.14.5
opencv-python==3.4.3.18
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.2
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
torch==0.4.1
torch-nightly==1.0.0.dev20181006
torchsummary==1.4
torchvision==0.2.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
